implemented sorcerer's ultimate, changed health type to float

+++ Assets/Scripts/Items/FallingPlatform.cs
@@ -12,7 +12,7 @@ public class FallingPlatform : MonoBehaviour {
     {
         float facing = collision.transform.position.x.CompareTo(0); // -1 if <0, 0 if ==0 and 1 if >0
         // les dommages seront proportionnels à la hauteur de chute
-        int damage = (int) (height - collision.transform.position.y);
+        float damage =  (height - collision.transform.position.y);
         Vector2 recoil = new Vector2(facing * 10, 2);
         Champion champion = collision.GetComponent<Champion>();
